feat(singledb): Conf runtime parameters to the current db conn session
