+++ frappe/database/__init__.py
@@ -6,16 +6,6 @@
 
 from __future__ import unicode_literals
 
-def init_db_session(func):
-	"""Configure runtime parameters scoped to the current database connection session.
-	"""
-	def wrapper(*args, **kwargs):
-		import frappe
-		conn = func(*args, **kwargs)
-		if frappe.conf.db_type == 'postgres' and frappe.local.tenant_id:
-			conn.sql(f"SELECT set_config('app.current_tenant', '{frappe.local.tenant_id}', false);")
-		return conn
-	return wrapper
 
 def setup_database(force, source_sql=None, verbose=None, no_mariadb_socket=False):
 	import frappe
@@ -34,7 +24,6 @@ def drop_user_and_database(db_name, root_login=None, root_password=None):
 		import frappe.database.mariadb.setup_db
 		return frappe.database.mariadb.setup_db.drop_user_and_database(db_name, root_login, root_password)
 
-@init_db_session
 def get_db(host=None, user=None, password=None, port=None):
 	import frappe
 	if frappe.conf.db_type == 'postgres':
